Dict Method
Python dictionary method program that uses the keys(), values(), and/or items() dict methods to find statistics about the student_grades dictionary. Find the following:

Print the name and grade percentage of the student with the highest total of points
Find the average score of each assignment.
Find and apply a curve to each student's total score, such that the best student has 100% of the total points.